requirements.txt
# Core Data Processing
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# Machine Learning
scikit-learn>=1.1.0
xgboost>=1.6.0
lightgbm>=3.3.0
imbalanced-learn>=0.9.0

# Deep Learning
tensorflow>=2.10.0
keras>=2.10.0

# Graph Analysis
networkx>=2.8.0
python-louvain>=0.16

# Natural Language Processing
nltk>=3.7
textblob>=0.17.1
gensim>=4.2.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
graphviz>=0.20.0

# Explainability
shap>=0.41.0

# Large Data Processing
dask>=2022.8.0

# Web Interface
streamlit>=1.12.0

# Report Generation
reportlab>=3.6.0
fpdf2>=2.6.0

# Configuration
pyyaml>=6.0

# API and Web
requests>=2.28.0

# Utilities
tqdm>=4.64.0
joblib>=1.1.0
python-dateutil>=2.8.0
pytz>=2022.1

# Testing
pytest>=7.1.0
pytest-cov>=3.0.0

# Development
jupyter>=1.0.0
ipykernel>=6.15.0